backend/pyproject.toml
[tool.poetry]
name = "openticket"
version = "0.1.0"
description = "Backend for Alan Finances"
authors = ["Ryan Downing <ryansdowning1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.7"
gunicorn = "20.1.0"
djangorestframework = "3.14.0"
gevent = "22.10.2"
psycopg2 = "2.9.5"
psycopg2-binary = "2.9.5"
django-extensions = "3.2.1"
django-filter = "22.1"
loguru = "0.6.0"
django-cors-headers = "3.14.0"

[tool.poetry.dev-dependencies]
pre-commit = "3.2.0"
ruff = "0.0.257"
black = "23.1.0"
ssort = "0.11.6"
ipykernel = "6.22.0"
PyYAML = "6.0"
uritemplate = "4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
